pyproject.toml
[project]
name = "rss_email"
version = "0.1.0"
description = "Delivers your daily RSS news via email."
authors = [{name = "Eamon Mason", email = "eamon.mason@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.28.49",
    "feedparser>=6.0.10",
    "PyRSS2Gen>=1.1",
    "beautifulsoup4>=4.12.2",
    "pydantic>=2.10.6",
    "anthropic>=0.52.0",
    "regex>=2024.11.6",
    "brotli>=1.1.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "flake8>=6.1.0",
    "prospector>=0.12.2",
    "pytest>=7.4.3",
    "moto>=4.2.7",
    "pylint>=3.0.2",
    "python-dotenv>=0.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.flake8]
max-line-length = 120

[tool.pytest.ini_options]
pythonpath = "src"

[tool.pylint.MASTER]
init-hook = "import sys; sys.path.insert(0, 'src')"

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "moto>=5.1.6",
    "prospector>=1.17.2",
    "pylint>=3.3.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
